Apogee: Prevent VC++ warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10920 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Apogee/Apogee.cpp
@@ -1225,7 +1225,7 @@ int CApogeeCamera::GetAltaAdGain( long & gain )
     }
     else
     {
-        gain = ApgCam->GainSixteenBit;
+        gain = static_cast<long>(ApgCam->GainSixteenBit);
     }
 
      return DEVICE_OK;
